refactor(authentication): Reorganize authentication-related packages
Move authentication controllers and services to dedicated `com.example.authentication.rest` and `com.example.authentication.services` packages respectively. Also move common utilities like `ExceptionHandler` and `MailService` directly under `com.example.common`. Update all affected package declarations and imports, including test files.
