+++ app/src/main/kotlin/com/example/authentication/services/UserService.kt
@@ -1,5 +1,6 @@
-package com.example.users
+package com.example.authentication.services
 
+import com.example.authentication.rest.CreateOrUpdateUserDTO
 import com.example.common.entities.DomainUser
 import com.example.common.repositories.UserRepository
 import jakarta.validation.constraints.NotBlank
