+++ app/src/test/kotlin/com/example/authentication/AccountIT.kt
@@ -1,11 +1,12 @@
-package com.example.users
+package com.example.authentication
 
+import com.example.authentication.SignInIT.Companion.DUMMY_EMAIL
+import com.example.authentication.SignInIT.Companion.DUMMY_PASSWORD
+import com.example.authentication.SignInIT.Companion.DUMMY_USERNAME
 import com.example.common.JPATestConfiguration
+import com.example.common.entities.DomainUser
 import com.example.common.events.PasswordChangedEvent
 import com.example.common.repositories.UserRepository
-import com.example.users.SignInIT.Companion.DUMMY_EMAIL
-import com.example.users.SignInIT.Companion.DUMMY_PASSWORD
-import com.example.users.SignInIT.Companion.DUMMY_USERNAME
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
