+++ app/src/test/kotlin/com/example/authentication/UsersIT.kt
@@ -1,11 +1,11 @@
-package com.example.users
+package com.example.authentication
 
+import com.example.authentication.SignInIT.Companion.DUMMY_EMAIL
+import com.example.authentication.SignInIT.Companion.DUMMY_PASSWORD
+import com.example.authentication.SignInIT.Companion.DUMMY_USERNAME
 import com.example.common.JPATestConfiguration
 import com.example.common.entities.DomainUser
 import com.example.common.repositories.UserRepository
-import com.example.users.SignInIT.Companion.DUMMY_EMAIL
-import com.example.users.SignInIT.Companion.DUMMY_PASSWORD
-import com.example.users.SignInIT.Companion.DUMMY_USERNAME
 import com.fasterxml.jackson.databind.ObjectMapper
 import java.util.stream.Stream
 import org.junit.jupiter.api.BeforeEach
